rename zarr objects used only for benchmarking
