+++ src/photon_mosaic/core/__init__.py
@@ -3,4 +3,4 @@ from .baserois import BaseRois
 from .arrayimaging import ArrayImaging, NumpyImagingEpoch, DaskImagingEpoch, NumpyRois
 from .generators import generate_random_imaging, generate_rois
 from .binaryimaging import read_binary
-from .zarrimaging import ZarrImaging, ZarrImagingEpoch, ZarrImagingEpochDask, ZarrImagingEpochNative
+from .zarrimaging_laura import ZarrImaging, ZarrImagingEpoch, ZarrImagingEpochDask, ZarrImagingEpochNative
